requirements.txt
langchain
langchain-core
langchain-community
python-dotenv
groq
pydantic
